rutas primer commit
