+++ client/src/components/Landing/Landing.jsx
@@ -3,11 +3,9 @@ import Login from "../Log in/Login";
 
 export default function Landing() {
   return (
-    <Fragment>
       <div>
         <h1>Blockbuster Henry</h1>
         <Login/>
       </div>
-    </Fragment>
   );
 }
