+++ client/src/components/Log in/Login.jsx
@@ -1,17 +1,23 @@
 import React,{useState} from 'react'
 import { useDispatch } from 'react-redux';
+import { Link } from 'react-router-dom';
+
+
 
 
 export default function Login  ()  {
+    
     const dispatch = useDispatch()
     const [user,setUser] = useState()
     function handleOnChange(e){
         setUser(e.target.value)
     }
-    function handleOnSubmit(){
+    function handleOnSubmit(e){
         e.preventDefault(e)
         dispatch('dispatch del login')//corregir
     }
+
+
   return (
     <div>
         <input
